update user data
