+++ app/components/homepage/blog/blog-card.jsx
@@ -1,5 +1,5 @@
 // @flow strict
-import { timeConverter } from '@/utils/timeConverter';
+import { timeConverter } from '@/utils/time-converter';
 import Image from 'next/image';
 import Link from 'next/link';
 import { BsHeartFill } from 'react-icons/bs';
