+++ app/components/homepage/footer.jsx
@@ -8,7 +8,7 @@ function Footer() {
     <div className="relative border-t bg-[#0d1224] border-[#353951]">
       <div className="mx-auto px-6 sm:px-12 lg:max-w-[70rem] xl:max-w-[76rem] 2xl:max-w-[92rem] py-6 lg:py-10">
         <div className="flex justify-center -z-40">
-          <div class="absolute top-0 h-[1px] w-1/2  bg-gradient-to-r from-transparent via-violet-500 to-transparent"></div>
+          <div className="absolute top-0 h-[1px] w-1/2  bg-gradient-to-r from-transparent via-violet-500 to-transparent"></div>
         </div>
         <div className="flex items-center justify-between">
           <p className="text-sm">
