+++ app/components/homepage/hero-section.jsx
@@ -1,5 +1,6 @@
 // @flow strict
 
+import { personalData } from "@/utils/data/personal-data";
 import Image from "next/image";
 import Link from "next/link";
 import { BsGithub, BsLinkedin } from "react-icons/bs";
@@ -24,39 +25,59 @@ function HeroSection() {
           <h1 className="text-3xl font-bold leading-10 text-white md:font-extrabold lg:text-[2.6rem] lg:leading-[3.5rem]">
             Hello, <br />
             This is {' '}
-            <span className=" text-pink-500">ABU SAID</span>
+            <span className=" text-pink-500">{personalData.name}</span>
             {` , a Professional `}
             <span className=" text-[#16f2b3]">Full-Stack</span>
             {' '} Software Developer.
           </h1>
 
           <div className="my-12 flex items-center gap-5">
-            <Link className="transition-all text-pink-500 hover:scale-125 duration-300" href='/'>
+            <Link
+              href={personalData.github}
+              target='_blank'
+              className="transition-all text-pink-500 hover:scale-125 duration-300"
+            >
               <BsGithub size={30} />
             </Link>
-            <Link className="transition-all text-pink-500  hover:scale-125 duration-300" href='/'>
+            <Link
+              href={personalData.linkedIn}
+              target='_blank'
+              className="transition-all text-pink-500 hover:scale-125 duration-300"
+            >
               <BsLinkedin size={30} />
             </Link>
-            <Link className="transition-all text-pink-500 hover:scale-125 duration-300" href='/'>
+            <Link
+              href={personalData.facebook}
+              target='_blank'
+              className="transition-all text-pink-500 hover:scale-125 duration-300"
+            >
               <FaFacebook size={30} />
             </Link>
-            <Link className="transition-all text-pink-500 hover:scale-125 duration-300" href='/'>
+            <Link
+              href={personalData.leetcode}
+              target='_blank'
+              className="transition-all text-pink-500 hover:scale-125 duration-300"
+            >
               <SiLeetcode size={30} />
             </Link>
-            <Link className="transition-all text-pink-500 hover:scale-125 duration-300" href='/'>
+            <Link
+              href={personalData.twitter}
+              target='_blank'
+              className="transition-all text-pink-500 hover:scale-125 duration-300"
+            >
               <FaTwitterSquare size={30} />
             </Link>
           </div>
 
           <div className="flex items-center gap-3">
-            <Link href="/courses" className="bg-gradient-to-r to-pink-500 from-violet-600 p-[1px] rounded-full transition-all duration-300 hover:from-pink-500 hover:to-violet-600">
+            <Link href="#contact" className="bg-gradient-to-r to-pink-500 from-violet-600 p-[1px] rounded-full transition-all duration-300 hover:from-pink-500 hover:to-violet-600">
               <button className="px-3 text-xs md:px-8 py-3 md:py-4 bg-[#0d1224] rounded-full border-none text-center md:text-sm font-medium uppercase tracking-wider text-[#ffff] no-underline transition-all duration-200 ease-out  md:font-semibold flex items-center gap-1 hover:gap-3">
                 <span>Contact me</span>
                 <RiContactsFill size={16} />
               </button>
             </Link>
 
-            <Link className="flex items-center gap-1 hover:gap-3 rounded-full bg-gradient-to-r from-pink-500 to-violet-600 px-3 md:px-8 py-3 md:py-4 text-center text-xs md:text-sm font-medium uppercase tracking-wider text-white no-underline transition-all duration-200 ease-out hover:text-white hover:no-underline md:font-semibold" role="button" href="/courses"
+            <Link className="flex items-center gap-1 hover:gap-3 rounded-full bg-gradient-to-r from-pink-500 to-violet-600 px-3 md:px-8 py-3 md:py-4 text-center text-xs md:text-sm font-medium uppercase tracking-wider text-white no-underline transition-all duration-200 ease-out hover:text-white hover:no-underline md:font-semibold" role="button" target="_blank" href={personalData.resume}
             >
               <span>Get Resume</span>
               <MdDownload size={16} />
