+++ app/components/homepage/skills.jsx
@@ -1,7 +1,7 @@
 // @flow strict
 
 import { skillsData } from "@/utils/data/skills";
-import { skillsImage } from "@/utils/skillImage";
+import { skillsImage } from "@/utils/skill-image";
 import Image from "next/image";
 import Marquee from "react-fast-marquee";
