+++ app/page.js
@@ -1,3 +1,4 @@
+import { personalData } from "@/utils/data/personal-data";
 import AboutSection from "./components/homepage/about-section";
 import Blog from "./components/homepage/blog";
 import ContactSection from "./components/homepage/contact-section";
@@ -8,7 +9,7 @@ import Projects from "./components/homepage/projects";
 import Skills from "./components/homepage/skills";
 
 async function getData() {
-  const res = await fetch(`https://dev.to/api/articles?username=said7388`)
+  const res = await fetch(`https://dev.to/api/articles?username=${personalData.devUsername}`)
 
   if (!res.ok) {
     throw new Error('Failed to fetch data')
