+++ utils/data/personal-data.js
@@ -1,0 +1,16 @@
+export const personalData = {
+  name: "ABU SAID",
+  designation: "Full-Stack Software Developer",
+  description: "My name is ABU SAID. I am a professional and enthusiastic programmer in my daily life. I am a quick learner with a self-learning attitude. I love to learn and explore new technologies and am passionate about problem-solving. I love almost all the stacks of web application development and love to make the web more open to the world. My core skill is based on JavaScript and I love to do most of the things using JavaScript. I am available for any kind of job opportunity that suits my skills and interests.",
+  email: 'abusaid7388@gmail.com',
+  phone: '+8801608797655',
+  address: 'Middle Badda, Dhaka, Bangladesh - 1212 ',
+  github: 'https://github.com/said7388',
+  facebook: 'https://www.facebook.com/abusaid.riyaz/',
+  linkedIn: 'https://www.linkedin.com/in/abu-said-bd/',
+  twitter: 'https://twitter.com/said7388',
+  stackOverflow: 'https://stackoverflow.com/users/16840768/abu-said',
+  leetcode: "https://leetcode.com/said3812/",
+  devUsername: "said7388",
+  resume: "https://drive.google.com/file/d/1eyutpKFFhJ9X-qpQGKhUNnVRkB5Wer00/view?usp=sharing"
+}
